Extracted common baseclass
