+++ community/security/src/main/java/org/neo4j/server/security/auth/AbstractUserRepository.java
@@ -1,0 +1,166 @@
+/*
+ * Copyright (c) 2002-2016 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.server.security.auth;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+import org.neo4j.kernel.lifecycle.LifecycleAdapter;
+import org.neo4j.server.security.auth.exception.ConcurrentModificationException;
+
+public abstract class AbstractUserRepository extends LifecycleAdapter implements UserRepository
+{
+    /** Quick lookup of users by name */
+    protected final Map<String, User> usersByName = new ConcurrentHashMap<>();
+
+    /** Master list of users */
+    protected volatile List<User> users = new ArrayList<>();
+
+    private final Pattern usernamePattern = Pattern.compile( "^[a-zA-Z0-9_]+$" );
+
+    @Override
+    public User getUserByName( String username )
+    {
+        return usersByName.get( username );
+    }
+
+    @Override
+    public void create( User user ) throws IllegalArgumentException, IOException
+    {
+        if ( !isValidUsername( user.name() ) )
+        {
+            throw new IllegalArgumentException( "'" + user.name() + "' is not a valid user name." );
+        }
+
+        synchronized (this)
+        {
+            // Check for existing user
+            for ( User other : users )
+            {
+                if ( other.name().equals( user.name() ) )
+                {
+                    throw new IllegalArgumentException( "The specified user already exists" );
+                }
+            }
+
+            users.add( user );
+
+            saveUsers();
+
+            usersByName.put( user.name(), user );
+        }
+    }
+
+    public abstract void saveUsers() throws IOException;
+
+    @Override
+    public void update( User existingUser, User updatedUser ) throws ConcurrentModificationException, IOException
+    {
+        // Assert input is ok
+        if ( !existingUser.name().equals( updatedUser.name() ) )
+        {
+            throw new IllegalArgumentException( "updatedUser has a different name" );
+        }
+
+        synchronized (this)
+        {
+            // Copy-on-write for the users list
+            List<User> newUsers = new ArrayList<>();
+            boolean foundUser = false;
+            for ( User other : users )
+            {
+                if ( other.equals( existingUser ) )
+                {
+                    foundUser = true;
+                    newUsers.add( updatedUser );
+                } else
+                {
+                    newUsers.add( other );
+                }
+            }
+
+            if ( !foundUser )
+            {
+                throw new ConcurrentModificationException();
+            }
+
+            users = newUsers;
+
+            saveUsers();
+
+            usersByName.put( updatedUser.name(), updatedUser );
+        }
+    }
+
+    @Override
+    public boolean delete( User user ) throws IOException
+    {
+        boolean foundUser = false;
+        synchronized (this)
+        {
+            // Copy-on-write for the users list
+            List<User> newUsers = new ArrayList<>();
+            for ( User other : users )
+            {
+                if ( other.name().equals( user.name() ) )
+                {
+                    foundUser = true;
+                } else
+                {
+                    newUsers.add( other );
+                }
+            }
+
+            if ( foundUser )
+            {
+                users = newUsers;
+
+                saveUsers();
+
+                usersByName.remove( user.name() );
+            }
+        }
+        return foundUser;
+    }
+
+    @Override
+    public int numberOfUsers()
+    {
+        return users.size();
+    }
+
+    @Override
+    public boolean isValidUsername( String username )
+    {
+        return usernamePattern.matcher( username ).matches();
+    }
+
+    @Override
+    public Set<String> getAllUsernames()
+    {
+        return users.stream().map( User::name ).collect( Collectors.toSet() );
+    }
+}
