+++ enterprise/security/src/test/java/org/neo4j/server/security/enterprise/auth/AuthTestBase.java
@@ -51,7 +51,6 @@ abstract class AuthTestBase<S>
     protected String READ_OPS_NOT_ALLOWED = "Read operations are not allowed";
     protected String WRITE_OPS_NOT_ALLOWED = "Write operations are not allowed";
     protected String SCHEMA_OPS_NOT_ALLOWED = "Schema operations are not allowed";
-    protected boolean HAS_ILLEGAL_ARGS_CHECK = false;
     protected boolean IS_EMBEDDED = true;
 
     protected String pwdReqErrMsg( String errMsg )
